vanna-repo/requirements.txt
vanna[ollama]
mysql-connector-python
flask
pandas
